services/rtpi-gpu-monitor/scripts/requirements.txt
flask==2.3.3
flask-cors==4.0.0
psutil==5.9.5
pyyaml==6.0
requests==2.31.0
prometheus-client==0.17.1
